# internal/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "VATUSAs internal API for use by other internal VATUSA services.",
        "title": "VATUSA Internal",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "VATUSA Support",
            "url": "http://www.swagger.io/support",
            "email": "vatusa6@vatusa.net"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.1"
    },
    "host": "api.vatusa.local",
    "basePath": "/internal/v1",
    "paths": {
...
        "/roster": {
            "get": {
                "description": "List rosters",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "roster"
                ],
                "summary": "List rosters",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/roster.Response"
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new roster",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "roster"
                ],
                "summary": "Create a new roster",
                "parameters": [
                    {
                        "description": "Roster",
                        "name": "roster",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/roster.Request"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/roster.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrResponse"
                        }
                    }
                }
            }
        },
        "/roster-request": {
            "get": {
                "description": "List all roster requests",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "roster-request"
                ],
                "summary": "List all roster requests",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/roster_request.Response"
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new roster request",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "roster-request"
                ],
                "summary": "Create a new roster request",
                "parameters": [
                    {
                        "description": "Roster Request",
                        "name": "roster_request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/roster_request.Request"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/roster_request.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrResponse"
                        }
                    }
                }
            }
        },
        "/roster-request/{id}": {
            "get": {
                "description": "Get a roster request",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "roster-request"
                ],
                "summary": "Get a roster request",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Roster Request ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/roster_request.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Update a roster request",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "roster-request"
                ],
                "summary": "Update a roster request",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Roster Request ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Roster Request",
                        "name": "roster_request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/roster_request.Request"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/roster_request.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a roster request",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "roster-request"
                ],
                "summary": "Delete a roster request",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Roster Request ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrResponse"
                        }
                    }
                }
            }
        },
        "/roster/{id}": {
            "get": {
                "description": "Get a roster",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "roster"
                ],
                "summary": "Get a roster",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Roster ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/roster.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Update a roster",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "roster"
                ],
                "summary": "Update a roster",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Roster ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Roster",
                        "name": "roster",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/roster.Request"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a roster",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "roster"
                ],
                "summary": "Delete a roster",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Roster ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "roster.Request": {
            "type": "object",
            "required": [
                "cid",
                "facility",
                "operating_initials",
                "status"
            ],
            "properties": {
                "cid": {
                    "type": "integer",
                    "example": 1293257
                },
                "facility": {
                    "type": "string",
                    "example": "ZDV"
                },
                "home": {
                    "type": "boolean",
                    "example": true
                },
                "instructor": {
                    "type": "boolean",
                    "example": false
                },
                "mentor": {
                    "type": "boolean",
                    "example": false
                },
                "operating_initials": {
                    "type": "string",
                    "example": "RP"
                },
                "status": {
                    "description": "Active, LOA",
                    "type": "string",
                    "enum": [
                        "active",
                        "loa"
                    ],
                    "example": "Active"
                },
                "visiting": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "roster.Response": {
            "type": "object",
            "properties": {
                "cid": {
                    "type": "integer",
                    "example": 1293257
                },
                "created_at": {
                    "type": "string",
                    "example": "2021-01-01T00:00:00Z"
                },
                "deleted_at": {
                    "description": "Soft Deletes for logging",
                    "type": "string",
                    "example": "2021-01-01T00:00:00Z"
                },
                "facility": {
                    "type": "string",
                    "example": "ZDV"
                },
                "home": {
                    "type": "boolean",
                    "example": true
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "instructor": {
                    "type": "boolean",
                    "example": false
                },
                "mentor": {
                    "type": "boolean",
                    "example": false
                },
                "operating_initials": {
                    "type": "string",
                    "example": "RP"
                },
                "status": {
                    "description": "Active, LOA",
                    "type": "string",
                    "example": "Active"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2021-01-01T00:00:00Z"
                },
                "visiting": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "roster_request.Request": {
            "type": "object",
            "required": [
                "cid",
                "reason",
                "request_type",
                "requested_facility",
                "status"
            ],
            "properties": {
                "cid": {
                    "type": "integer",
                    "example": 1293257
                },
                "reason": {
                    "type": "string",
                    "example": "I want to transfer to ZDV"
                },
                "request_type": {
                    "enum": [
                        "visiting",
                        "transferring"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/types.RequestType"
                        }
                    ],
                    "example": "visiting"
                },
                "requested_facility": {
                    "type": "string",
                    "example": "ZDV"
                },
                "status": {
                    "enum": [
                        "pending",
                        "accepted",
                        "rejected"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/types.StatusType"
                        }
                    ],
                    "example": "pending"
                }
            }
        },
        "roster_request.Response": {
            "type": "object",
            "properties": {
                "cid": {
                    "type": "integer",
                    "example": 1293257
                },
                "created_at": {
                    "type": "string",
                    "example": "2021-01-01T00:00:00Z"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "reason": {
                    "type": "string",
                    "example": "I want to transfer to ZDV"
                },
                "requestType": {
                    "$ref": "#/definitions/types.RequestType"
                },
                "requested_facility": {
                    "type": "string",
                    "example": "ZDV"
                },
                "status": {
                    "$ref": "#/definitions/types.StatusType"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2021-01-01T00:00:00Z"
                }
            }
        },
...
        "types.RequestType": {
            "type": "string",
            "enum": [
                "visiting",
                "transferring"
            ],
            "x-enum-varnames": [
                "Visiting",
                "Transferring"
            ]
        },
        "types.StatusType": {
            "type": "string",
            "enum": [
                "pending",
                "accepted",
                "rejected"
            ],
            "x-enum-varnames": [
                "Pending",
                "Accepted",
                "Rejected"
            ]
        },
...
        "utils.ErrResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "application-specific error code",
                    "type": "integer"
                },
                "error": {
                    "description": "application-level error message, for debugging",
                    "type": "string"
                },
                "status": {
                    "description": "user-level status message",
                    "type": "string"
                }
            }
        }
    }
}